Refactor out account data
