+++ src/app/components/accounts/account-summary/account-summary.ts
@@ -1,11 +1,10 @@
 import {Component, input, computed, inject} from '@angular/core';
-import {Account, Category} from 'ynab';
 
 import {CurrencyCopyButton} from '../../common/currency-copy-button/currency-copy-button';
 import {Currency} from '../../common/currency/currency';
 import {DropdownButton, ButtonTheme} from '../../common/dropdown-button/dropdown-button';
 import {YnabStorage} from '../../../../lib/ynab/ynab_storage';
-import {SettingsStorage} from '../../../../lib/firebase/settings_storage';
+import {AccountAllocation} from '../../../../lib/accounts/account_data';
 
 @Component({
   selector: 'ya-account-summary',
@@ -20,18 +19,8 @@ import {SettingsStorage} from '../../../../lib/firebase/settings_storage';
 export class AccountSummary {
   readonly account = input.required<AccountAllocation>();
 
-  protected readonly allocatedAmount = computed<number>(() => {
-    let sum = 0;
-    for (const category of this.account().categories) {
-      // We always take a positive value here because negative values still need
-      // to be covered somehow.
-      sum += Math.abs(category.balance);
-    }
-    return sum;
-  });
-
   protected readonly delta = computed<number>(() => {
-    return this.account().account.cleared_balance - this.allocatedAmount();
+    return this.account().account.cleared_balance - this.account().total;
   });
 
   protected readonly hasAllocations = computed<boolean>(() => {
@@ -52,12 +41,4 @@ export class AccountSummary {
   });
 
   private readonly ynabStorage = inject(YnabStorage);
-}
-
-/**
- * The special data structure that this component renders.
- */
-export interface AccountAllocation {
-  readonly account: Account;
-  readonly categories: Category[];
 }
